Release v1.0.11a
This is a source-only prerelease of GixSQL. It is a consolidated version that incorporates the latest fixes in the internal work repository and is meant as a stepping point for testing before v1.0.12, which still requires some work. No updated Changelog is provided, you will have to look at the issue list on GitHub.